修复 GPU 材质解析与可见性管线黑屏
- 修复可见像素分类 WGSL 保留字导致的 Shader 编译失败\n- 清理 Material ABI v4 中残留的 material_id 访问\n- 补齐材质纹理索引的显式数值转换\n- 修正 Material Resolve 顶点阶段的可见性 Bind Group 权限\n- 统一 GPU Counter 禁用路径的 512B ABI 大小\n- 增加 Shader ABI、Bind Group 和计数器回归测试

+++ OEngine/src/shaders/packed_material_resolve.ts
@@ -179,10 +179,10 @@ fn read_uv_direct(geometry: GpuGeometryRecord, uv_set: u32, vertex: u32) -> vec2
     );
   }
   if format == ${GPU_UV_FORMAT.Unorm8x2}u {
-    return vec2f(read_u8(offset), read_u8(offset + 1u)) / 255.0;
+    return vec2f(f32(read_u8(offset)), f32(read_u8(offset + 1u))) / 255.0;
   }
   if format == ${GPU_UV_FORMAT.Unorm16x2}u {
-    return vec2f(read_u16(offset), read_u16(offset + 2u)) / 65535.0;
+    return vec2f(f32(read_u16(offset)), f32(read_u16(offset + 2u))) / 65535.0;
   }
   return vec2f(0.0);
 }
